Feat : search component type
